% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\docType{data}
\name{assessment}
\alias{assessment}
\title{Assessment}
\format{
A data frame with 4 rows and 3 variables:
\describe{
\item{class}{Numeric class}
\item{assessment}{Assessment}
\item{action}{Proposed action}
}
}
\usage{
assessment
}
\description{
Data frame containing the assessment and actions based on observed and
expected
}
\keyword{datasets}
