% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{consistency}
\alias{consistency}
\title{Assess Consistency}
\usage{
consistency(data)
}
\arguments{
\item{data}{Dataframe}
}
\value{
Dataframe
}
\description{
Assess if WHPT scores are consistent with expected classification.
}
\examples{
predictions <- whpt_predict(demo_data)
data <- inner_join(demo_data, predictions, by = c("sample_id" = "sample_id"))
assessments <- consistency(data)
}
