% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whpt.R
\name{whpt}
\alias{whpt}
\title{Predict WHPT}
\usage{
whpt(data)
}
\arguments{
\item{data}{Dataframe of GIS based predictors with 20 variables:
\describe{
\item{Location code}{Sample ID - unique identifer for sample}
\item{NGR}{National Grid Reference - Great Britain only}
\item{Date}{Date as character class in 2012-12-31 format only}
\item{Sampled date}{Predicted total number of scoring WHPT families}
\item{SX}{Coordinated where GIS predictors come from}
\item{SY}{Coordinated where GIS predictors come from}
\item{EX}{Coordinated where GIS predictors queried}
\item{EY}{Coordinated where GIS predictors queried}
\item{Altitude}{Altitude in metres}
\item{d_f_source}{Distance from source in metres}
\item{logaltbar}{Log altitude in metres of catchment upstream}
\item{log_area}{Log area of catchment upstream in km squared}
\item{disch_cat}{Discharge category}
\item{slope}{Slope in m / km}
\item{chalk}{Proporation of chalk in catchment}
\item{clay}{Proporation of clay in catchment}
\item{hardrock}{Proporation of hardrock in catchment}
\item{limestone}{Proporation of limestone in catchment}
\item{peat}{Proporation of peat in catchment}
\item{shape_Length}{Length of the river section represented in GIS layer}
}}
}
\value{
Dataframe consisting of three columns sample_id, ASPT and NTAXA
\describe{
\item{sample_id}{Sample ID - unique identifer for sample}
\item{ASPT}{Predicted Average Score Per Taxa (WHPT Distinct families)}
\item{NTAXA}{Predicted total number of scoring WHPT families}
}
}
\description{
Script to run prediction for WHPT ASPT and NTAXA Based on whpt-metric-model -
using random forest / GIS predictors only based on RIVPACS reference dataset
See \code{demo_data} for data structure / predictors required Predictors
come from EA website:
https://environment.data.gov.uk/DefraDataDownload/?mapService=EA/RICT&mode=spatial
}
\examples{
whpt_predictions <- whpt(demo_data)
}
