% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whpt_app.R
\name{whpt_app}
\alias{whpt_app}
\title{Open whpt assessment web app}
\usage{
whpt_app()
}
\description{
Open whpt assessment an interactive shiny web app.
}
\details{
\code{whpt_app()} opens interactive shiny app.

Using the app:
\enumerate{
\item Select your location and add observed aspt, ntaxa.
\item View assessment and download results.
\item There is no third step.
}
}
\examples{
\dontrun{
whpt_app()
}

}
