% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whpt-predict.R
\name{whpt_predict}
\alias{whpt_predict}
\title{Predict WHPT}
\usage{
whpt_predict(data)
}
\arguments{
\item{data}{Data frame of \code{date_taken} and GIS based predictors with 14
variables. Variable names can be lower or upper-case, and in any order.
Extra variables can be present, these variables are ignored.
\describe{
\item{sample_id}{Sample ID - unique identifer for sample}
\item{date_taken}{Date as character class in \code{2012-12-31} format only}
\item{ngr}{National Grid Reference - Great Britain only in \verb{NT 00990 65767}
format only}
\item{altitude}{Altitude in metres}
\item{d_f_source}{Distance from source in metres}
\item{logaltbar}{Log altitude in metres of catchment upstream}
\item{log_area}{Log area of catchment upstream in km squared}
\item{disch_cat}{Discharge category}
\item{slope}{Slope in m / km}
\item{chalk}{Proporation of chalk in catchment}
\item{clay}{Proporation of clay in catchment}
\item{hardrock}{Proporation of hardrock in catchment}
\item{limestone}{Proporation of limestone in catchment}
\item{peat}{Proporation of peat in catchment}
}}
}
\value{
Dataframe consisting of three variables sample_id, \code{index} and
\code{predicted_response}.
\describe{
\item{sample_id}{Sample ID - unique identifer for sample}
\item{index}{Either \verb{Reference ASPT} or \verb{Reference NTAXA})}
\item{predicted_response}{Predicted response value}
}
}
\description{
Run prediction for WHPT ASPT and NTAXA Based on whpt-metric-model -
using random forest / GIS predictors only based on RIVPACS reference dataset.
See \code{demo_data} for data structure / predictors required. Predictors
come from EA website:
https://environment.data.gov.uk/DefraDataDownload/?mapService=EA/RICT&mode=spatial
}
\examples{
predictions <- whpt_predict(demo_data)
}
