% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whpt_scores_model.R
\docType{data}
\name{whpt_scores_model}
\alias{whpt_scores_model}
\title{WHPT prediction model}
\format{
A data frame with 2 rows and 3 variables:
\describe{
\item{DETERMINAND}{Value which is being predicted, either \code{ASPT} or \code{NTAXA}}
\item{recipe}{This the 'recipe' to use to scale, center and transform data before applying model}
\item{model_final}{This is the random forest model for predicting \code{ASPT} or \code{NTAXA}}
}
}
\usage{
whpt_scores_model
}
\description{
A data frame containing two models (ASPT and NTAXA). Contains a 'recipe' for
transforming data and the model to run.
}
\keyword{datasets}
