% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whpts.R
\name{whpts}
\alias{whpts}
\title{Assess WHPT Consistency}
\usage{
whpts(data)
}
\arguments{
\item{data}{Dataframe}
}
\value{
Dataframe
}
\description{
Assess if WHPT scores are consistent with expected classification.
}
\examples{
results <- whpts(demo_data)
}
