% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whpt_app.R
\name{whpt_app}
\alias{whpt_app}
\title{Open whpt assessment web app}
\usage{
whpt_app()
}
\description{
Open whpt assessment an interactive shiny web app.
}
\details{
\code{whpt_app()} opens interactive shiny app.

Using the app:
\enumerate{
\item Select your file (or download a template to fill out)
\item Or select your location and add data, aspt, ntaxa.
\item View and download results.
\item
}
}
\examples{
\dontrun{
whpt_app()
}

}
